Add files via upload
Data Preprocessing challenges. Untreated datasets are proposed at the beginning of the challenges in order to process them and remove outliers and other anomalies within the dataset for being statistically ready for the training and prediction of the model. It is mainly used pandas and numpy libraries along with pre-processing techniques based on Python